Clean up API
Move SystemReserved and KubeReserved into KubeletConfiguration struct
Convert int64 to int32 for some external type fields so they match internal ones
tLS* to tls* for JSON field names
Fix dependency on removed options.AutoDetectCloudProvider
Change floats in KubeletConfiguration API to ints

+++ contrib/mesos/pkg/executor/service/service.go
@@ -190,7 +190,7 @@ func (s *KubeletExecutorServer) runKubelet(
 	}
 
 	// make a separate client for events
-	eventClientConfig.QPS = s.EventRecordQPS
+	eventClientConfig.QPS = float32(s.EventRecordQPS)
 	eventClientConfig.Burst = int(s.EventBurst)
 	kcfg.EventClient, err = clientset.NewForConfig(eventClientConfig)
 	if err != nil {
